Add config.py for configuration and work on CAN-FIX device listing
The config.py file will be used for general program configuration
for now.  It may change to something else in the future but this
is quick and dirty.  Continued work on the device list module.
The device list will be the central repository of information for
configuration of the CAN-FIX devices.  Sorta like an electronic
data sheet