Add netsim program and start work on network CAN adapter
The netsim program will be used instead of the simulation adapter for
simulating CAN traffic to the Utility.  This gives more control without
having to modify the actual Utility program further.  The network adapter
will be used to communicate with this simulator and could also be used
to communicate to a CAN<->TCP/IP server somewhere.