Remove the need for response from the network CAN server
The serial port adapters wait for responses from the devices after
they send commands to set bit rate, open channels or send frames.
Since the network adapter is using TCP this seems unnecessary.  TCP
is a reliable protocol.  If the other end doesn't get the command we'll
get an error.  No need to duplicate that in our code.